chartable: fix zoom window on Wayland
On Wayland, popup windows are for menu type widgets and thus the
protocol requires a device grab, which gucharmap doesn't do.

As a result, the zoom window is dismissed instantly on Wayland.

For the use case here, best would be to use a subsurface for the zoom
window, which can be obtained by "attaching" the popup to the parent so
that gtk+ can select the subsurface type for the popup.

Bugzilla: https://bugzilla.gnome.org/show_bug.cgi?id=759229