Make the build reproducible
The Reproducible Builds effort [0] noticed that gucharmap could not be
built reproducibly.

This is because it includes the absolute build path in the generated
gucharmap-type-builtins.h file:

- /* enumerations from "/build/1st/gucharmap-13.0.1/obj-x86_64-linux-gnu/../gucharmap/gucharmap-charmap.h" */
+ /* enumerations from "/build/2/gucharmap-13.0.1/2nd/obj-x86_64-linux-gnu/../gucharmap/gucharmap-charmap.h" */

This commit also makes the parallel change to the .c file; although this
does not make a difference to the build, it feels like the right thing
to do.

[0] https://reproducible-builds.org/

Resolves: https://gitlab.gnome.org/GNOME/gucharmap/-/issues/38
Bug-Debian: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=955341