lib: Enable alphabetical sorting of the Unicode blocks
Sort the tree view column and not directly the model.

Add a string containing the block start to the block chapters model and
sort using this by default.

Each time the column header of the Unicode blocks is clicked, the column
to be sorted will switch between block start and block name, but nothing
but the block name column will ever be displayed.

https://bugzilla.gnome.org/show_bug.cgi?id=620180
Fixes: https://gitlab.gnome.org/GNOME/gucharmap/-/issues/35